requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
bcrypt==5.0.0
cffi==1.17.1
click==8.1.8
cryptography==46.0.7
dnspython==2.6.1
ecdsa==0.19.2
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.124.4
h11==0.16.0
idna==3.11
mongoengine==0.29.3
motor==3.6.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.23
pydantic==2.10.6
pydantic-core==2.27.2
pymongo==4.9.2
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
typing-extensions==4.13.2
uvicorn==0.33.0
